Update main.py to fix post dict becoming nested
I was using https://github.com/ILogOutOnTheToilet/kemono-dl but today (Nov 25 2024) I noticed it had broken.  Based on the error message and debugging I was able to implement a basic workaround in the clean_post() function in main.py:  un-nest the new post variable with 'post = post['post']' . This restores the kind of dict structure that the rest of the code clearly expected to see.